+++ datatune/__init__.py
@@ -1,7 +1,5 @@
 from datatune.agent.agent import Agent
 from datatune.core.filter import filter
 from datatune.core.map import map
-from datatune.core.op import finalize
-from datatune.core.ibis.map_ibis import map_Ibis
-from datatune.core.ibis.filter_ibis import filter_Ibis
-__all__ = ["map", "filter", "finalize", "Agent", "map_Ibis", "filter_Ibis"]
+from datatune.core.dask.op import finalize
+__all__ = ["map", "filter", "finalize", "Agent"]
